+++ Examples/PostEffects/Application.cpp
@@ -2,7 +2,6 @@
 
 #include <IncludesGl.h>
 #include <Utils/RandomTools.h>
-#include <Utils/IO.h>
 
 #include "imgui.h"
 #include "imgui_impl_sdl2.h"
