+++ Examples/Text/Application.cpp
@@ -1,7 +1,6 @@
 #include "Application.h"
 
 #include <Utils/RandomTools.h>
-#include <Utils/IO.h>
 
 #include "imgui.h"
 #include "imgui_impl_sdl2.h"
